feat: adjust styling, content, responsiveness
This commit corrects the styling responsible for responsiveness for the
web app. There is also new content added, adjusting the about section to
collapse the personal story, and add experience to the section. The
'reach me' has been moved to the start of the page and adjusted to
contain a Resume pdf link. More projects have been added.